requirements.txt
apache-beam==2.52.0
attrs==23.1.0
beam-postgres-connector==0.1.3
certifi==2023.11.17
charset-normalizer==3.3.2
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
dnspython==2.4.2
docopt==0.6.2
dunamai==1.19.0
fastavro==1.9.1
fasteners==0.19
grpcio==1.60.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.6
Jinja2==3.1.2
Js2Py==0.74
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
MarkupSafe==2.1.3
numpy==1.24.4
objsize==0.6.1
orjson==3.9.10
packaging==23.2
poetry-dynamic-versioning==0.24.0
proto-plus==1.23.0
protobuf==4.25.1
psycopg2-binary==2.9.9
pyarrow==11.0.0
pyarrow-hotfix==0.6
pydot==1.4.2
pyjsparser==2.7.1
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rpds-py==0.13.2
six==1.16.0
tomlkit==0.12.3
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
zstandard==0.22.0
